# openapispec.yml
openapi: 3.1.0
...
info:
  title: openspecapi
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
  /ingredients/find:
    get:
      summary: Your GET endpoint
      tags: []
      operationId: get-ingredients-find
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    createdAt:
                      type: string
                    updatedAt:
                      type: string
                    ingredientName:
                      type: string
                    ingredientDescription:
                      type: string
                    ingredientPictureLink:
                      type: string
                x-examples:
                  Example 1:
                    - id: 5cb36bb8-85f5-44c1-b77e-36c961cd1f0b
                      createdAt: '2022-11-25T11:38:04.000Z'
                      updatedAt: '2022-11-25T11:38:04.000Z'
                      ingredientName: chicken
                      ingredientDescription: one nice chicken
                      ingredientPictureLink: www.chicken.de
                    - id: eab11593-268a-4410-b1a2-a121a55111e8
                      createdAt: '2022-11-25T11:38:32.000Z'
                      updatedAt: '2022-11-25T11:38:32.000Z'
                      ingredientName: egg
                      ingredientDescription: L Egg
                      ingredientPictureLink: www.egg-l.de
              examples:
                Example 1:
                  value:
                    - id: 5cb36bb8-85f5-44c1-b77e-36c961cd1f0b
                      createdAt: '2022-11-25T11:38:04.000Z'
                      updatedAt: '2022-11-25T11:38:04.000Z'
                      ingredientName: chicken
                      ingredientDescription: one nice chicken
                      ingredientPictureLink: www.chicken.de
                    - id: eab11593-268a-4410-b1a2-a121a55111e8
                      createdAt: '2022-11-25T11:38:32.000Z'
                      updatedAt: '2022-11-25T11:38:32.000Z'
                      ingredientName: egg
                      ingredientDescription: L Egg
                      ingredientPictureLink: www.egg-l.de
        '400':
          description: Bad Request
  '/ingredients/findByName/{name}':
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    createdAt:
                      type: string
                    updatedAt:
                      type: string
                    ingredientName:
                      type: string
                    ingredientDescription:
                      type: string
                    ingredientPictureLink:
                      type: string
                x-examples:
                  Example 1:
                    - id: eab11593-268a-4410-b1a2-a121a55111e8
                      createdAt: '2022-11-25T11:38:32.000Z'
                      updatedAt: '2022-11-25T11:38:32.000Z'
                      ingredientName: egg
                      ingredientDescription: L Egg
                      ingredientPictureLink: www.egg-l.de
                    - id: 5c5e0fd8-22dd-481b-a437-e38cf9780731
                      createdAt: '2022-11-25T11:38:24.000Z'
                      updatedAt: '2022-11-25T11:44:40.000Z'
                      ingredientName: egg
                      ingredientDescription: M Egg
                      ingredientPictureLink: www.egg-m.de
        '400':
          description: Bad Request
      operationId: get-ingredients-findByName
      parameters: []
      description: ''
    parameters:
      - schema:
          type: string
        name: name
        in: path
        required: true
  /ingredients/create:
    post:
      summary: ''
      operationId: post-ingredients-create
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                  ingredientName:
                    type: string
                  ingredientDescription:
                    type: string
                  ingredientPictureLink:
                    type: string
                x-examples:
                  Example 1:
                    id: 0d087968-1cf7-4821-98b3-ad68a3e77afe
                    createdAt: '2022-11-26T16:00:46.675Z'
                    updatedAt: '2022-11-26T16:00:46.675Z'
                    ingredientName: stoplighstudio
                    ingredientDescription: atopligh food
                    ingredientPictureLink: www.stop.de
              examples:
                Example 1:
                  value:
                    id: string
                    createdAt: string
                    updatedAt: string
                    ingredientName: string
                    ingredientDescription: string
                    ingredientPictureLink: string
        '400':
          description: Bad Request
        '404':
          description: Not Found
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ingredientName:
                  type: string
                ingredientDescription:
                  type: string
                ingredientPictureLink:
                  type: string
              x-examples:
                Example 1:
                  ingredientName: stoplighstudio
                  ingredientDescription: atopligh food
                  ingredientPictureLink: www.stop.de
            examples:
              Example 1:
                value:
                  ingredientName: stoplighstudio
                  ingredientDescription: atopligh food
                  ingredientPictureLink: www.stop.de
  '/ingredients/ingredientUpdate/{id}':
    put:
      summary: ''
      operationId: put-ingredients-ingredientUpdate
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                  ingredientName:
                    type: string
                  ingredientDescription:
                    type: string
                  ingredientPictureLink:
                    type: string
                x-examples:
                  Example 1:
                    id: 5cb36bb8-85f5-44c1-b77e-36c961cd1f0b
                    createdAt: '2022-11-25T11:38:04.000Z'
                    updatedAt: '2022-11-26T16:42:57.969Z'
                    ingredientName: chicken
                    ingredientDescription: one nice chicken
                    ingredientPictureLink: www.chicken.de
        '400':
          description: Bad Request
        '404':
          description: Not Found
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              x-examples:
                Example 1:
                  ingredientName: test
                  ingredientDescription: test
                  ingredientPictureLink: test
              properties:
                ingredientName:
                  type: string
                ingredientDescription:
                  type: string
                ingredientPictureLink:
                  type: string
      description: ''
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
  '/ingredients/ingredientDelete/{id}':
    delete:
      summary: ''
      operationId: delete-ingredients-ingredientDelete
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
        '403':
          description: Forbidden
      parameters: []
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
...